Add deserializing to ForkliftSerializer; move it to the connectors package
