+++ connectors/kafka/src/test/java/forklift/connectors/KafkaConnectorSerializerTests.java
@@ -1,8 +1,8 @@
 package forklift.connectors;
 
 import forklift.Forklift;
+import forklift.connectors.ForkliftSerializer;
 import forklift.integration.server.TestServiceManager;
-import forklift.producers.ForkliftSerializer;
 import forklift.source.sources.TopicSource;
 
 import org.junit.AfterClass;
