+++ plugins/replay/src/main/java/forklift/replay/ReplayES.java
@@ -3,6 +3,7 @@ package forklift.replay;
 import forklift.Forklift;
 import forklift.connectors.ForkliftConnectorI;
 import forklift.connectors.ForkliftMessage;
+import forklift.connectors.ForkliftSerializer;
 import forklift.consumer.Consumer;
 import forklift.consumer.ConsumerService;
 import forklift.consumer.ConsumerThread;
@@ -14,7 +15,6 @@ import forklift.decorators.LifeCycle;
 import forklift.decorators.Service;
 import forklift.message.Header;
 import forklift.producers.ForkliftProducerI;
-import forklift.producers.ForkliftSerializer;
 import forklift.producers.ProducerException;
 import forklift.source.ActionSource;
 import forklift.source.SourceI;
